add transfer widget in in-memory data
